refactor: move zipfile to single file
